requirements.txt
yt-dlp>=2023.12.30
python-telegram-bot>=20.0
curl_cffi>=0.5.10
